python: Select processes to profile based on "python*" in their exe's basename
This is more accurate than "pgrep python", because a Python script exec'd directly won't
have "python" in its comm (hence the different PYTHON_KNOWN_PROCS we pgrep'd for).
This new approach is more accurate, and it can be taken to further to actually finding
processes with ".*python.*" in their /proc/pid/maps file.