python: Select processes to profile based on python/libpython in their maps
In 3e9381bc4ff7 I suggested this change as an improvement, but it's actually required
because we currently miss executables which load libpython, such as uwsgi.

This change means that we'll prefer the Python samples from *any* process which has
libpython loaded, even if it doesn't even run Python code now. We might need to address
that later, by making smarter decisions on the source for our samples.

Fixes: 3e9381bc4ff7