java: Clean up container-side temporary directory inside profile_process()
We're encountering a race where, if profiling is stopped abruptly with ^C, the main
thread raises KeyboardInterrupt which calls JavaProfiler.__exit__(), removing the
temp dirs. Meanwhile, in the profiling thread, the _stop_event triggers and we
execute "jattach stop,..." but meanwhile the temporary directory (containing the log file,
output file etc) might be removed.