java: Use plain /tmp/gprofiler as the tmpdir in different namespaces
Currently we don't clean up the main "temporary directory" of different namespaces
(see previous commit in this PR for the reason). This means that many empty
temporary directories will remain if we are to use different name for the base
directory every time.

This way, the only directory left eventually is /tmp/gprofiler.