perf: Use /tmp/gprofiler/perf-buildids as the buildid cache for perf
Avoid cluttering other parts of the filesystem (by default, it uses /root/.debug).
When gprofiler runs as a container, it doesn't really matter (both /root/ and /tmp
are just part of the container filesystem).
But we plan to allow running gprofiler as an executable, and in that case, we
should keep our files in /tmp/gprofiler.