tests: More fine-grained testing for specific runtimes or profilers. (#19)
The `profiler` fixture assumes only a single profiler implementation for a
particular runtime. This change inlines it and requires unit tests to be
written for a particular implementation. The existing test for the gprofiler
docker image remains the same. In addition, we now no longer execute test
applications of runtimes that are not being tested, and we don't execute an
application both within a container and on the host for the same test
permutation.

After this change, it is easier to write tests for a particular profiler
implementation, or for a particular application runtime.