Bundle gProfiler to one executable
gProfiler bundled to one executable using Pyinstaller, here are several main points:
The build is on to of Centos 7 docker image to support a wide range of machines.
Pytest test added to test the executable (CI is configured to test it in various linuxes versions).
GH actions added which are build, test and deploy the executable to GH releases.

Bundle to one executable will extend the deployment options of gProfiler
and allow simple and frictionless deployment in more environments.