Integrate Python eBPF profiler (#36)
This commit integrates PyPerf into gProfiler. See https://github.com/Granulate/bcc/pull/1.

PyPerf is an eBPF-based Python profiler, which profiles all Python processes across the system.
When its preconditions are met, PyPerf will be used instead of py-spy. If PyPerf can not be used (e.g due to
an old kernel version not supporting our required eBPF features), py-spy will be used as the fallback instead.

Along the way, gProfiler was modified to support truly "continuous" mode of operation: PyPerf is started *once*
when gProfiler is invoked, and remains running until gProfiler is stopped. Every profiling interval, PyPerf is polled to
dump its latest collected stacks. This is done because PyPerf's bootstrap is heavier (compared to perf & py-spy, ...) so it
is best if we invoke it just once.
In the future we'll convert the existing profilers (perf & async-profiler) to use the continuous mode.

Co-authored-by: Yonatan Goldschmidt <yonatan.goldschmidt@granulate.io>