PyInstaller: Improvements & robustness (#45)
* PyInstaller: Dockerfile: Fail the build if any of gprofiler modules are not found by Analysis

* PyInstaller: Install UPX so it's used during build

Helps reducing size by a bit. We already have it enabled in the spec file.

* PyInstaller: Dockerfile: Avoid "COPY . ."

"COPY . ." is never cached :/ there's always some change.

* PyInstaller: staticx: Included libraries needed by our resources

And fail the build if any is missing

* PyInstaller: Dockerfile: Build PyPerf on centos:7

Building it on ubuntu:20.04 results in incompatible libraries expected by PyPerf vs. found
on centos:7.
There's also the libc issue (we want the old libc for older kernels) - but it's not a real
issue because PyPerf won't run on old kernels anyway.

* python: PyPerf: Under staticx, run PyPerf with the packaged ld.so

* PyInstaller: Dockerfile: Don't manually specify libgcc

* python: PyPerf: Use .staticx.interp instead of hardcoding ld-2.17.so

The former is a link to it

* scripts/list_needed_libs.sh: Use echo instead of cat

* scripts/list_needed_libs.sh: Capture output in quotations