java: Do not readlink("/proc/pid/exe")
There's no need to move the the target PID namespace.

The link can be passed to execve itself. It has the benefit of running even if
the file was deleted (for example - if the Java installation was upgraded then
readlink("/proc/pid/exe") returns the path with " (deleted)", but executing
/proc/pid/exe will run the "old" version just fine (as long as all libs exist, etc).

This matches the fix I posted for py-spy: https://github.com/benfred/py-spy/pull/364