Pin image versions in build Dockerfiles
Stuff like 816549e reminded me that we should take the same actions in our Dockerfiles,
especially those that relate to the build outputs, which need to remain consistent and reproducible
for the same commit in this repository.

This is obviously not enough, we still run many apt-get/yum operations which depend on external
resources.. But it's a start.