Shrink Docker image size: 789 MB -> 328 MB
Started with dive (https://github.com/wagoodman/dive) but eventually there were
2 simple actions that drastically increased the size, so I avoided them:
1. Installing too many packages due to recommendation
2. Generation of egg files & copying the resources to multiple directories

This is without squashing. --squash currently reduces only 1 MB, because we don't
remove anything in any layer. If we do get to do that, we can add squashing.

Size can be decreased even further, 328 MB is not the lowest we can get to.
Next steps should be trying alpine, IMO.