Require all runtime profilers to get "comm" on their own
We need them to get comm since we want the final result to appear the same,
also when running in '--perf-mode none'.
Instead of making it "configurable", I just decided that they should always return
comm; this simplifies the code.

Additionally, this solves for once and for all what we tried to solve in
6201bb0fa6a97d by upgrading perf - now, we'll *always* get the process' comm,
and not the thread's, no matter what.