Use /proc/pid/comm as the first stack of process profiles (#165)
psutil.Process.name() seems to get it from /proc/pid/stat, but it returns
very long strings (over the 16 bytes of comm) so it doesn't make sense. It seems like
its return value is the first entry from /proc/pid/cmdline.

Anyway, we'll just get /proc/pid/comm to be accurate with what "perf" gets. Our "perf"
gets the "exec comm" (which is the comm initially set for the process upon exec), or the comm
of the main thread (for processes that are already alive).