Build async-profiler on CentOS 7
My trigger for that are the lousy mirrors of CentOS 6, which make our CI flaky.
In any case, it's a hassle to continue using CentOS 6 in any way.

This commit fixes the build to be CentOS 7 based. It's old enough glibc to run on
old kernels, and I verified that the only relevant symbol which required glibc>2.12
now gets versioned to something older as well.