ProfileData API (#326)
The idea is to use a more rich output format for profilers - not just PID -> stackcollapse, but PID -> object with numerous fields, one of them being stackcollapse.

Currently, the 2 additional fields are the appid and app metadata. Logically, it makes sense that the runtime profilers retrieve them, and not the generic area in gprofiler - because these are per-runtime anyway. Code is structured more "sensibly" when runtime profilers return these values.

Also, it serves a technical purpose: if profilers collect & attach the appid & app metadata directly to the profile data, we can be sure that:
* App id & metadata are always collected consistently - if a certain app has "appid: X", then all stacks of it will have that appid. Previously it wasn't the case, because we could profile a process that'd go down before we had the chance to get its app id (was retrieved post profiling).
* No chance for PID reuse collision - PID being reused by another app and we get the wrong data.

That's ^^ not true yet for pyperf, because pyperf still attaches this data post profiling. This will change in a PR after #281 - where we will move the process selection to gprofiler from pyperf, and then gprofiler will take the metadata & id of apps before pyperf begins to profile them.