Profile spawning processes (#281)
This PR aims to solve the "short processes" issue: gProfiler runs profiling sessions of 60 seconds (that's the default). Upon each session, the different profilers (JavaProfiler, PySpyProfiler etc) scan through all currently running processes and invoke their respective profilers for each relevant process.

This works nicely for processes running for long durations.However, it fails miserably in environments where processes are spawning and exiting quickly. For example, for an app with average lifetime of 10 seconds, we'll only see profiling data from the respective profiler for less than 1/6 of the time! (because on average, we'll catch only one out of 6 processes, and that process might already be halfway through its 10 seconds...). The situation was even worse for Java before #265, because we didn't read the output file if the process exited before the profiling session has ended. So by design - processes shorter than 60 seconds were not profiled... (py-spy & rb-spy do not suffer from this problem; PyPerf & phpspy, by design, do not suffer from it).

This PR add --profile-spawned-processes, which makes use of proc_events to catch new processes as they spawn and start profiling them with the respective profiler. Once we gain enough confidence in it, we will enable it by default.

Co-authored-by: Tomer Yogev <tomer.yogev@granulate.io>