java: Fix async-profiler compatibility on CentOS 6 (#453)
This PR fixes compatibility with CentOS 6. When we switched to build on CentOS 7, we didn't pass the -static-libstdc++ flag, and at the point when we fixed that regression, we didn't test for CentOS 6 support, which broke because -static-libstdc++ pulls in some symbols which require newer glibc.

This PR avoids -static-libstdc++ on glibc; it's required for musl-based because musl-based JVMs aren't always dependent on libstdc++; and we've never seen that with glibc.

Additionally, this PR adds a test for final dependency resolving compatibility on CentOS 6.

Also included is a cherry pick of https://github.com/jvm-profiling-tools/async-profiler/commit/b8c8db45d7ba2adfc99a21fb1b561002a8500a0c, which prevents async-profiler from crashing if it is re-loaded after a previous load has failed.