tests: Turn command_line fixture lazy
dotnet_command_line & java_command_line are 2 fixtures that require java/dotnet to be installed,
and if they're not installed, the fixture fails. By being eager - this fails the entire test
suite. Now they're lazy so the lack of java/dotnet will fail only relevant tests.

Additionally, I made the fixture cached - to avoid re-compilation of the artifacts. This will
be improved in the future by avoiding the build at all and copying the artifacts from the matching container.

Related: #481