Introduce and reporting DSO names in perf and java-ap stack profiles (#491)
This implements what was discussed in https://github.com/Granulate/gprofiler/issues/432 for perf and async-profiler.

This change formalizes the way we display file names (we chose inside parenthesis, to match what we already have for Python).
Then, we add --insert-dso-name - currently not enabled by default but will be - to add the DSO name in parenthesis for all frames, including frames where we did manage to symbolicate. This helps understanding which libraries are used.