metrics: Send samples per series instead of per metric name (#724)
Currently samples are only generated once per metric name, ignoring additional labels which might result in additional series. This PR standardizes the data classes to be used for samples and the entire snapshot, and fixes the code to report samples for all series even if they have the same metric name but different labels.